added check for valid screen names

+++ build_researcher_twitter_lists.py
@@ -1,30 +1,30 @@
-import tweepy
 import os
-import pandas as pd
-from pandas import json_normalize
 import json
 from pathlib import Path
 import time
+from datetime import datetime
 
+import tweepy
+import pandas as pd
 from cvedata.acknowledgements import get_researcher_twitter_map_json
 
+# Setup twitter client
+
+# set tokens
 import auth_tweepy
 
-auth = tweepy.OAuthHandler(os.getenv("consumer_key"),
-                           os.getenv("consumer_secret"))
-auth.set_access_token(os.getenv("access_token"),
-                      os.getenv("access_token_secret"))
-
+auth = tweepy.OAuthHandler(os.getenv("consumer_key"),os.getenv("consumer_secret"))
+auth.set_access_token(os.getenv("access_token"),os.getenv("access_token_secret"))
 
 api = tweepy.API(auth, wait_on_rate_limit=True)
-client = tweepy.Client(
-    bearer_token=os.getenv("bearer_token"),
-    consumer_key=os.getenv("consumer_key"),
-    consumer_secret=os.getenv("consumer_secret"),
-    access_token=os.getenv("access_token"),
-    access_token_secret=os.getenv("access_token_secret"),
-    wait_on_rate_limit=True,
-)
+# client = tweepy.Client(
+#     bearer_token=os.getenv("bearer_token"),
+#     consumer_key=os.getenv("consumer_key"),
+#     consumer_secret=os.getenv("consumer_secret"),
+#     access_token=os.getenv("access_token"),
+#     access_token_secret=os.getenv("access_token_secret"),
+#     wait_on_rate_limit=True,
+# )
 
 # Used to create list
 # list_name = "researchers with a CVE"
@@ -38,7 +38,6 @@ list_id = 1582329731005005825
 researchers_json = get_researcher_twitter_map_json()
 screen_name_to_researcher_map = {
     str(v[0]).lower(): k for k, v in researchers_json.items()}
-
 
 screen_names = set()
 for researcher in researchers_json:
@@ -76,11 +75,41 @@ print(f"All screen names len {len(screen_names)}")
 screen_names_to_add = screen_names.difference(current_screen_names)
 screen_names_to_add = list(screen_names_to_add)
 print(f"Screen names to add len {len(screen_names_to_add)}")
-sub_size = 3
 
-print(f"Need to add {len(screen_names_to_add)} researchers")
+# check validity of names
+valid_screen_names = []
+sub_size = 100
+
+invalid_sn_path = Path('invalid_screen_names.json')
+
+if invalid_sn_path.exists() and datetime.fromtimestamp(invalid_sn_path.stat().st_mtime).day == datetime.now().day:
+    # load cached version
+    invalid_sn = json.loads(invalid_sn_path.read_text())
+else:
+    # check names again
+    for i in range(0, len(screen_names_to_add), sub_size):
+        sub_screen_names = [sn for sn in screen_names_to_add[i: i + sub_size]]
+        users = api.lookup_users(screen_name=sub_screen_names)
+        
+        for user in users:
+            if user.screen_name:
+                valid_screen_names.append(user.screen_name.lower())
+
+    invalid_sn = set(screen_names_to_add).difference(set(valid_screen_names))
+    invalid_sn = list(invalid_sn)
+    invalid_sn_path.write_text(json.dumps(invalid_sn))
+
+
+
+print(f"Invalid Screennames {len(invalid_sn)}")
+
+# remove invalid screen names (suspended, doesn't exist, etc..)
+screen_names_to_add = list(set(screen_names_to_add).difference(set(invalid_sn)))
+
+print(f"Need to add {len(screen_names_to_add)} valid researchers")
 print(f"Starting member count = {len(current_screen_names)}")
 
+sub_size = 10 # twitter doesn't like you to add 100 at once
 member_count = len(current_screen_names)
 for i in range(0, len(screen_names_to_add), sub_size):
     sub_screen_names = [sn for sn in screen_names_to_add[i: i + sub_size]]
@@ -93,12 +122,13 @@ for i in range(0, len(screen_names_to_add), sub_size):
         resp = api.add_list_members(
             list_id=list_id, screen_name=sub_screen_names)
         new_member_count = resp.member_count
+        added = new_member_count - member_count
         print(
-            f"Added {new_member_count - member_count} of {len(sub_screen_names)} attempted")
+            f"Added {added} of {len(sub_screen_names)} attempted")
         member_count = new_member_count
 
         # delete current members as file is not outdated
-        if current_members_path.exists():
+        if current_members_path.exists() and added > 0:
             current_members_path.unlink()
 
         time.sleep(1)
